+++ data/src/main/java/ru/sogya/projects/activityandcharity/data/mapper/AchievementsMapper.kt
@@ -1,0 +1,7 @@
+package ru.sogya.projects.activityandcharity.data.mapper
+
+import ru.sogya.projects.activityandcharity.data.model.AchievementsData
+import ru.sogya.projects.activityandcharity.domain.model.AchievementsDomain
+
+fun AchievementsDomain.toData(): AchievementsData =
+    AchievementsData(id, title, description, photoUrl)
