+++ data/src/main/java/ru/sogya/projects/activityandcharity/data/repository/firebase/UserRepositoryImpl.kt
@@ -1,0 +1,6 @@
+package ru.sogya.projects.activityandcharity.data.repository.firebase
+
+import ru.sogya.projects.activityandcharity.domain.repository.firebase.UserStoreRepository
+
+class UserRepositoryImpl : UserStoreRepository {
+}
