--- data/src/main/java/ru/sogya/projects/activityandcharity/model/request/CreateUserRequest.kt
@@ -1,16 +1,0 @@
-package ru.sogya.projects.activityandcharity.model.request
-
-import com.google.gson.annotations.SerializedName
-
-data class CreateUserRequest(
-    @SerializedName("name")
-    val name: String,
-    @SerializedName("email")
-    val email: String,
-    @SerializedName("departament")
-    val department: Int,
-    @SerializedName("type")
-    val type: Int,
-    @SerializedName("password")
-    val password: String
-)
