--- data/src/main/java/ru/sogya/projects/activityandcharity/model/response/AuthResponse.kt
@@ -1,8 +1,0 @@
-package ru.sogya.projects.activityandcharity.model.response
-
-import com.google.gson.annotations.SerializedName
-
-data class AuthResponse(
-    @SerializedName("auth_success")
-    val authSuccess: Boolean
-)
