--- data/src/main/java/ru/sogya/projects/activityandcharity/repository/firebase/UserRepositoryImpl.kt
@@ -1,6 +1,0 @@
-package ru.sogya.projects.activityandcharity.repository.firebase
-
-import ru.sogya.projects.activityandcharity.domain.repository.firebase.UserStoreRepository
-
-class UserRepositoryImpl : UserStoreRepository {
-}
