--- data/src/main/java/ru/sogya/projects/activityandcharity/util/Constants.kt
@@ -1,16 +1,0 @@
-package ru.sogya.projects.activityandcharity.util
-
-object Constants {
-    const val APP_NAME = "appName"
-    const val BASE_URL = "https://0rlwqbsl-5000.euw.devtunnels.ms/api"
-    const val USER_LOGIN = ""
-    const val USER_ID = "userId"
-    const val USER_PASSWORD = ""
-    const val IS_AUTH = "isAuth"
-    const val DATABASE_NAME = "sportivity_database"
-
-
-    //UI
-    const val KM = "km"
-    const val MIN = "min"
-}
